Move ExpectedIssuesParser & PhpTestFile
