+++ php-frontend/src/testFixtures/java/org/sonar/php/utils/PhpTestFile.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.plugins.php.api.tests;
+package org.sonar.php.utils;
 
 import java.io.File;
 import java.io.IOException;
